refactor(front): share the agent column and drop the ComponentProps indirection
Review follow-ups: spell out the icon component type instead of indexing
Icon's props, and move the agent cell into the shared column module so both
tables get the same column def and tooltip.
