+++ front/components/workspace/analytics/automations/automationsTriggerColumns.tsx
@@ -1,10 +1,15 @@
 import { getIcon } from "@app/components/resources/resources_icons";
 import type { TriggerRowData } from "@app/components/workspace/analytics/automations/AutomationsTriggersRowsTable";
-import { CreditsCell } from "@app/components/workspace/analytics/creditsTableCells";
+import {
+  AvatarNameCell,
+  CreditsCell,
+  EntityTooltipCard,
+} from "@app/components/workspace/analytics/creditsTableCells";
 import type { AutomationTriggerRow } from "@app/lib/api/analytics/automations/triggers";
 import { normalizeWebhookIcon } from "@app/lib/webhook_source";
 import { assertNeverAndIgnore } from "@app/types/shared/utils/assert_never";
 import {
+  Avatar,
   Button,
   ChevronDown,
   ChevronUp,
@@ -14,15 +19,14 @@ import {
   Tooltip,
 } from "@dust-tt/sparkle";
 import type { ColumnDef } from "@tanstack/react-table";
-import type { ComponentProps } from "react";
+import type { ComponentType } from "react";
 
-function TypeLabel({
-  visual,
-  label,
-}: {
-  visual: ComponentProps<typeof Icon>["visual"];
+interface TypeLabelProps {
+  visual: ComponentType<{ className?: string }>;
   label: string;
-}) {
+}
+
+function TypeLabel({ visual, label }: TypeLabelProps) {
   return (
     <Tooltip
       label={label}
@@ -36,7 +40,11 @@ function TypeLabel({
   );
 }
 
-export function TypeCell({ trigger }: { trigger: AutomationTriggerRow }) {
+interface TypeCellProps {
+  trigger: AutomationTriggerRow;
+}
+
+export function TypeCell({ trigger }: TypeCellProps) {
   switch (trigger.kind) {
     case "schedule":
       return (
@@ -81,6 +89,63 @@ export function nameColumn<T extends TriggerRowData>(): ColumnDef<T> {
         <span className="truncate text-sm font-semibold">
           {info.row.original.name}
         </span>
+      </DataTable.CellContent>
+    ),
+  };
+}
+
+interface AgentCellProps {
+  agent: AutomationTriggerRow["agent"];
+}
+
+function AgentCell({ agent }: AgentCellProps) {
+  const content = (
+    <div className="min-w-0">
+      <AvatarNameCell
+        name={agent.name}
+        imageUrl={agent.pictureUrl}
+        size="xxs"
+      />
+    </div>
+  );
+
+  if (!agent.description) {
+    return content;
+  }
+
+  return (
+    <Tooltip
+      label={
+        <EntityTooltipCard
+          avatar={
+            <Avatar
+              name={agent.name}
+              visual={agent.pictureUrl ?? undefined}
+              size="xs"
+            />
+          }
+          name={agent.name}
+          description={agent.description}
+          modelId={agent.modelId}
+          modelDisplayName={agent.modelDisplayName}
+        />
+      }
+      className="p-3"
+      tooltipTriggerAsChild
+      trigger={content}
+    />
+  );
+}
+
+export function agentColumn<T extends TriggerRowData>(): ColumnDef<T> {
+  return {
+    id: "agent",
+    header: "Agent",
+    enableSorting: false,
+    meta: { className: "w-44", headerAlign: "left" },
+    cell: (info) => (
+      <DataTable.CellContent className="w-full justify-start">
+        <AgentCell agent={info.row.original.agent} />
       </DataTable.CellContent>
     ),
   };
